+++ mdsine2/pylab/base/__init__.py
@@ -1,5 +1,4 @@
 from .errors import UndefinedError, MathError, InitializationError, InheritanceError, GraphIDError
-from .graph import BaseNode, DataNode, Node, Graph, isnode, isgraph, get_default_graph
+from .graph import BaseNode, DataNode, Node, Graph, isnode, isgraph, get_default_graph, Saveable, TraceableNode, issavable, istraceable
 from .random import israndom, seed, normal, lognormal, truncnormal, multivariate_normal, gamma, beta, \
     sics, invchisquared, invgamma, uniform, negative_binomial, bernoulli
-from .trace import Saveable, TraceableNode, issavable, istraceable
